feat: Tag overhaul (#23)
* feat: Tag overhaul

Allows for much better compatibility with other food mods and starts restructuring tags in order to follow the new 1.20.X conventions. Forge tags are needed for now, so the structure is still a bit wonky, but once neoforge replaces it most farm_and_charm: tags will be able to be replace with c: tags, except for f&c-specific tags.

Main requirement before merging is applying the same tag updates to other Farm&Charm-based mods - or at least do so before the next update

* fix: wrong tag namespace for c:crops elements

Doesn't break anything, but better for semantics

* Add farm_and_charm:strawberry tag